+++ src/main/java/org/example/Solution/day04/Day04Solver.java
@@ -11,27 +11,27 @@ import java.util.List;
 public class Day04Solver extends AbstractDayXXSolver {
 
     @Override
-    public Integer partOneSolution() {
+    public Long partOneSolution() {
         List<ScratchCard> cards = getScratchCards();
-        return cards.stream().map(ScratchCard::getScore).reduce(0, Integer::sum);
+        return cards.stream().map(ScratchCard::getScore).reduce(0L, Long::sum);
     }
 
     @Override
-    public Integer partTwoSolution(){
-        List<Integer> intersections = getScratchCards().stream()
+    public Long partTwoSolution(){
+        List<Long> intersections = getScratchCards().stream()
                 .map(ScratchCard::getIntersection)
                 .toList();
         return calculateCopies(intersections);
     }
 
-    private Integer calculateCopies(List<Integer> intersections) {
-        List<Integer> copies = new ArrayList<>(Collections.nCopies(intersections.size(), 1));
+    private Long calculateCopies(List<Long> intersections) {
+        List<Long> copies = new ArrayList<>(Collections.nCopies(intersections.size(), 1L));
         for (int i = 0; i < intersections.size(); i++){
             for (int j = 1; j <= intersections.get(i); j++){
                 copies.set(i + j, copies.get(i) + copies.get(i + j));
             }
         }
-        return copies.stream().reduce(0, Integer::sum);
+        return copies.stream().reduce(0L, Long::sum);
     }
 
     private List<ScratchCard> getScratchCards() {
